qunit: Remove custom QUnit 1.x setup()/teardown() aliases
From <https://api.qunitjs.com/QUnit/module/#changelog>, these
alises were deprecated in 2014 with QUnit 1.16, and removed in 2016
with the release of QUnit 2.0. Most of these were migrated en-mass
thanks to eslint-plugin-qunit, but those that used it via
`QUnit.newMwEnvironment()` were not detected and migrated. This has
been done over the past few weeks now (tracked at T250045).

There is more clean up to be done in this file, but this minimal
change removes only the aliasing functionality.

Bug: T250045
Depends-On: I890caad67604f1c018341877db8d7f3cffe916fc
Change-Id: I4200c94eff9eabea18567e02f79721a5529da80f