qunit: Micro-optim for testrunner $(document) and mw.log
Instead of re-creating $(document) and re-saving live references,
do this only once.

This also removes any chance of mw.log potentially getting lost or
out of sync, e.g. if it becomes a stub at run-time we might accidentally
save that at the time suppressWarnings() is called, which is now
naturally impossible.

Change-Id: I681de1a0b1407e47219151a5c32c7237ec5e3f68