specials: Use setTitle() on GET forms to strip subpage everywhere
When a form is displayed on a page like 'Special:PrefixIndex/Foo',
and you want the form submission to go to just 'Special:PrefixIndex'
and pass 'Foo' as a parameter, you need to strip it from the title.
We often did it using an ugly pattern where a DerivativeContext
was created with the new title, and used as the form's context.

It is actually my fault, I introduced it a very long time ago in
2e68e7e8b851464df85241167f657f839ca247e2, because I thought that
HTMLForm#setTitle should be deprecated. In hindsight, doing it this
way is super ugly, and there is no value in deprecating that method
(and the patch that would deprecate it was never merged).

Recently, change af4969a15058d6b9c12e568e6450847485125178 removed
this pattern in a few special pages and reminded me of it. This
change removes it everywhere else that I could find.

Change-Id: I360cb3ab0d088273b7dba3c711b91e71b9a3cc59