Call TempUserCreatedRedirect hook with the post-create user
The TempUserCreatedRedirect hook was called with the request user, which
is typically an IP address. This causes an exception in the proposed
CentralAuth patch, because User::newFromName() returns false for an IP
address.

So, pass the "user for save", which is the actual created user with a
correct user ID.

Change-Id: Id4f0e393216da4a1da35922e27f97050931761f0