mediawiki.util: Make internal repeatString() function non-public
Follows-up 63a952c. This function wasn't meant to be a public API
and is not likely to last on its own given no external use case,
easy to locally own where needed, and a built-in exists in ES6. In
order to avoid making this difficult to remove later on, keep it
private.

Also:
- remove branch for now-impossible Infinity condition.
- remove condition for impossible empty string,
  the loop achieves the same if this failure scenario were to happen
  by accident in the future.
- remove condition for impossible negative count,
  if a future accident causes this to be possible, it will now fail
  by effectively assuming count=0 via the loop whereas previously
  it effectively assumed count=1 by returning str as-is.
- further localise var declarations where possible per current code
  conventions.

Bug: T301174
Change-Id: Icd4633cc75973cd9e6be5f0f4fc25eed5cfef4b9