UserGroupManager: Fix dbDomain in addUserToGroup() deferred update
This fixes a fatal error in the deferred update from addUserToGroup(),
due to it not specifying a dbDomain parameter. This bug was
introduced in 2020 with commit 40b88d635be6c4c.

== Scenario

Use meta.wikimedia.org/wiki/Special:UserRights to act on
`Example@otherwiki`.

> DBConnectionError:
> Cannot access the database: Unknown database 'metawiki'

== Problem

The domain param to getConnection() is optional for the common case of
connecting to the core database of the wiki's main load balancer.

This is the one kind of shared state we tolerate since it is
deterministically derived from site configuration, and is safely
dependency-injected from service wiring.

But, when obtaining a LoadBalancer for a foreign wiki, one must
basically always pass this to getConnection() as otherwise it will
try to connect to the local wiki database (e.g. metawiki) while
calling the LoadBalancer for otherwiki.

== Consideration

* The LoadBalancer objects should and are shared for dbs on the
  same section, thus preserving the dbDomain from the LBFactory
  method call is impractical, and probably also too magical/fragile
  in practice due to increasing the distance between these
  responsiblities and creating more easy-to-forget state.

* There is no obvious default for any given loadbalancer other than
  the local wiki's db. There are typically many wikis on the same
  LB.

* For external LBs, we actually do sometimes make use the current wiki,
  e.g. ExternalStore is structured as tables inside a DB for each
  wiki.

Bug: T193565
Change-Id: Ic82c264f7d3bbb1403600d21cfed3516f854332a
(cherry picked from commit 8b8834739f5292bba10512797abd385817e17e2c)