logging: Add index hint when asking for a specific user
It seems optimizer can't pick the right index when the user has done
many log actions. See the ticket for more information.

There are two indexes for actor and they are not that different so even
the wrong pick is fine but to make it more performant, make sure you get
log_actor_type_time when you filter based on both actor and type.

Bug: T303089
Change-Id: Ie0e7b5cc3a245dc28157a0276f6cce62f90b797f