Add put() and delete() shorthands to rest.js
These methods are not uncommon in a REST framework, so it'd be useful to
have a shorthand for them.

Add QUnit tests, copying the one for post().

Note that the REST framework actually ignores the request body for
DELETE requests, but we may want to change that (T309388), and so it is
supported here.

Change-Id: I9a4372314be1785c2124be354bd4151d595b6260