Fix Tyap (kcg) namespace names
The wrong spellings are mostly my fault,
a find-and-replace error. This also fixes
the names for "Category" and "Special page",
that were translated incorrectly.

The first wiki in this language was only created
recently, and the wrong names weren't used anywhere,
so there's no reason to preserve them for
backwards compatibility.

Change-Id: I015e4c2cc4bb001e40312f662b4af9c4f8800483