phpunit: Use customized TestDox output
Switch PHPUnit to use the more verbose TestDox format, which allows
us to see exactly which tests have been executed. This is useful when
making changes to test configuration.

Use the MediaWikiPHPUnitResultPrinter to override the CliTestDoxPrinter
and use the actual class names / test method names, instead of the
prettified versions from TestDox, to improve greppability.

Bug: T297287
Change-Id: I9d8561ff868e6ff0a44d32120f2799b34d25e618