Show change tags on Special:MergeHistory
The order in the result row is build using a message,
which needs the new parameter in the english message text.

The stylesheet mediawiki.interface.helpers.style is needed to show
parenthesis around the tags. It also adds parenthesis to the "bytes",
makes the comment in italic and format revision deleted information in
strike-grey.

Bug: T309647
Change-Id: I1fb481e6e62372874e3c65ec5e4cebf17291217f