search: Hint the SearchEngine about the set of snippets to provide
Some SearchEngine's optimize their snippet generation to not overload
Special:Search with too many snippets, but that implementation detail
is difficult to document for a generic API and leads to confusion
about why a requested snippet may or may not be included in the
response.

Add a way for the api to hint to the SearchEngine what it expects from
the snippet generation. SearchEngine's are under no obligation to
respect this hint. Longer term this may evolve into an API that
provides a more concrete set of expectations around what types of
snippets are expected.

Bug: T306477
Related: I38d58a5af45eef7cd1798312af137a70f73c54c3
Change-Id: Ied664e3c1b7d246bca914f10c8fd2e36ff6418ec