ApiParse: When querying TOC sections, also return whether it should be hidden
The table of contents is supposed to be hidden when the page has less
than 4 headings, but it can be overridden by several magic words.
The API provided no way to find out if it should be shown or hidden
for skins that don't use the built-in table of contents.

Using getTOCHTML() to check this is weird, but the TOC code is
currently enduring a rewrite, and this is how it's done in OutputPage.

Bug: T294950
Bug: T310083
Change-Id: I6cf76c870124c162dc1bcbc2f7e9ca0c5fdcd10e