Clarify usage of $offset in UserTimeCorrection
The constructor param $offset was assigned to $this->offset, but this is
useless and confusing. parse() is called immediately, and it always
overrides $this->offset with a value derived from the $timeCorrection
string, *unless* the correction type is SYSTEM. So rename the parameter
and make it clear that it's only used for SYSTEM.

Also add a FIXME about offsets after "System|" are ignored. This looks
like a leftover from older MW versions, and it's already noted in
SetupDynamicConfig.php, too. However, I'm too scared to remove it,
because I fear it could break something (especially because I've no idea
what the values in the DB could be, considering the other FIXME in the
parse() docblock).

Change-Id: I0156d5f7824f87ba6bc7f395135a4e1c80c517f4