HTMLInfoField: Ensure ID is included for the 'info' field.
Ensure that the ID is included in the generated label field for
the 'info' type (which is defined in HTMLInfoField.php).
This is needed for a OOUI conversion in the CheckUser extension,
as a JS script relies on an ID being present for a element that
only the 'info' type suits.

Change-Id: If4724e6699aa3ec076817a7c4727cbff184b94e7