Guard CustomConvertCommand against placeholders within values
Because the str_replaces were done sequentially, it was possible that
the replacement values for the placeholders of the
CustomConvertCommand would insert placeholder values of their own.

Find all placeholders at once on the original value using
preg_replace_callback and make sure each is replaced only once.

Based on a suggestion by User:SacredSum

Bug: T308394
Change-Id: Id672996454445dd200e57cf207c56e3268a3fcc8