Use ParserOutputFlags::SHOW_TOC in OutputPage
Parser output flags are intended to be "set only".  That means that
extensions can *set* the SHOW_TOC flag, but once set SHOW_TOC can't be
reset, and the TOC will be shown in the result if *any* of the merged
ParserOutputs on the OutputPage contain the SHOW_TOC flag set.  That's
consistent with prior merge behavior.

There's a possibility extensions are playing other games with the
Table Of Contents (TOC) in the OutputPageParserOutput hook; be
compatible with prior behavior for now but we may need to provide
additional functionality (perhaps a way to merge new section data) if
needed.

Bug: T310083
Change-Id: I0909ac85c6c785d9089b077a16923c61d6a09996