Support requesting undetermined SVG language
The default for multilanguage SVGs is the content language. In order
to show element with no explicit language, you have to request the
'und' aka undetermined language.

Because this was not special cased, this was not accessible and would
always fallback to English.

Make the default language code a const.

Bug: T310235
Change-Id: I6941f46d176ddf3dc811b29d580b1c5b8856eb1c