SpecialMyLanguage: Add method for finding a Title for transclusion
Before this commit, the base page is always returned in some
situations, e.g. if the UI and content language are the same. This
avoids extra lookups and ensures a cleaner URL.

In some cases, we would prefer to get a subpage even in this
situation, e.g. transcluding a translatable page. Add a method
that always performs the lookup.

Also update EditPage::getPreloadedContent to use the new method.

Bug: T309329
Change-Id: I739cee94620bc8d0689a4d3ec5898bfd75c18b8a