Parsoid API: emit strong eTags
This patch changes the etag emitted for HTML in GET requests to be
marked as a "strong" (byte-by-byte) eTag. This matches the behavior of
RESTbase.

This also adds checks for strong ETags to the e2e tests for the v1/page
and v1/revision endpoints. This is done for completeness. These endpoints
do not use the code changed here, the tests would have passed without
the change to ParsoidHandler.

NOTE: This addresses the issue of weak ETags coming from parsoid
endpoints only for setups that do NOT use RESTbase and/or Varnish.
RESTbase never had the issue. Varnish apparently weakens any ETag when
applying compression.

Context: The eTags returned in response to GET requests to parsoid endpoints
are later used in an If-Match header in requests to the transform endpoint,
when transforming HTML to wikitext. The intent is for the request to
fail if the server doesn't have the "stash" entry identified by the
etag, which would be needed to get a clean conversion using SelSer.

For this to work, the etag must be "strong", that ism they must not have
the W/ prefix that indicates a weak prefix. The HTTP spec mandates that
If-Match headers using a weak etag must always cause the request to
fail.

Bug: T310710
Change-Id: I075b20f0937b51b7fcde2e9fa9ac1cec8c7eec87