page: Change visibility of some WikiPage members to protected
After surveying consumers of these members, apart from the WikiPage
class itself that uses it internally and FlagabbleWikiPage in the
FlaggedRevs ext which subclasses WikiPage, I don't see any other
consumers of this member.

So it can be safely marked as protected as we're sure nothing is
consuming this publicly.
See: https://codesearch.wmcloud.org/search/?q=%5C-%5C%3EmDataLoaded&i=nope&files=&excludeFiles=&repos=,
https://codesearch.wmcloud.org/search/?q=%5C-%5C%3EmPreparedEdit&i=nope&files=&excludeFiles=&repos=,
https://codesearch.wmcloud.org/search/?q=%5C-%5C%3EmLatest&i=nope&files=&excludeFiles=&repos=

NOTE: The tests in ArticleTest.php was removed as the WikiPage::mLatest
      is now a protected member and can't be accessed publicly.

TODO: The last of it's kind is WikiPage::mTitle which will be removed in
      a later patch and that one will be quite bit and tricky.

Change-Id: I78c5b4c31090fe2e01b4cc922ec77131be3223dc