Use ParserOptions::setSuppressTOC() for Javascript content
This ensures we don't show ToCs in vector-2022 for Javascript pages
which contain raw <h2> elements.

This version uses the same "canonical options" hack that is used for
language conversion tables.  These should both be replaced by a
(not yet existing) mechanism in the future: T313455 is the
task for that.

Bug: T307691
Depends-On: I35e199cca40c0e4359ac493e5806dcf4ae49321c
Change-Id: Iba6a8b6c59bf91e3d06896f0a610c3c3e52e6564