Declare dependency on symfony/polyfill-php73
We use a function that’s introduced in PHP 7.3 and polyfilled in this
package (array_key_first(), in rdbms), but previously didn’t declare a
dependency on the package. (We declared a dependency on polyfill-php80,
but that doesn’t transitively pull in polyfill-php73.) Let’s add it.

(The package is already part of mediawiki/vendor, using the same
version.)

Change-Id: Ia76d4945087ba51354e21136f36f2614cf7d9bf8