Rename parsoid stash backend config var.
This changes the name of a key in the ParsoidCacheConfig array:
StashType is more in line with other configuration variables that refer
to on aobject cache type.

This is not a breaking change since ParsoidCacheConfig was only
introduced recently and is not in used in WMF config nor present in any
release.

This is a Follow-up to Ic67dc43ed9843810e4b180127f9a3bb7608f7608.

Change-Id: I9e61a6f94ec8bf1716c7d735f9bb17af4461fc9c