ResourceLoader: replace FileModule::collateFilePathListByOption with private helper
All callers provide 'media' for $option and 'all' for $default, just use
those values directly. Class is not stable to extend so this protected
method is not part of the stable interface, codesearch shows no
code on gerrit outside of core that calls the method. Since the $option
parameter is removed, the name no longer makes sense, renamed and
made private.

Change-Id: I875942d5b8c339031fd302f547c9cf16328c9bf2