ParserCache: ensure we know a revision ID
ParserCache::checkOutdated relies on ParserOutput::getCacheRevisionId() to determine
whether a revision is still current after loading it from the cache. If
the revision ID is 0 or null, this will result in false negatives, and
the revision will always be considered outdated.

It is better to detect and report this before writing the ParserOutput to the cache.

This also adds an assertion in DerivedPageDataUpdater that will trigger
an exception if we try to write to the parser cache before the revision
has been saved and the ID is known.

Change-Id: I242b769afbc7e1ae1e3f218d451f04945dfa8be4