language: Add missing `@ingroup`, subgroup "Languages" and ungroup files
== Ungroup file blocks

Remove `@ingroup` from `@file` blocks and keep only the class block.

This matches similar changes previously applied to API, Skins, Profile,
and ResourceLoader.

This helps make the API documentation easier to navigate.
E.g. Modules -> Language in the sidebar of
<https://doc.wikimedia.org/mediawiki-core/master/php/> as well as
<https://doc.wikimedia.org/mediawiki-core/master/php/group__Language.html>

These are currently cluttered with tons of duplicate entries for files
and classes both. We only need to group files that aren't also
documented as a class (e.g. message files, entry points, other scripts
or files that we mainly consider a data file). This has the helpful
side-effect that we don't encourage duplication of the class
description (or worse, place useful docs only in the file block), and
makes the class files consistently start with a mentally ignorable
block. Basically, unless there's something other than a class, don't
describe or group the file itself.

== Missing group

Various classes in this subtree were missing the `Language` group,
or were using different group from before T225756.

== Subgroup

For ease of navigation, move Converter subclasses to a group called
"Languages", which for documentation purposes is a subgroup of
"Language". The next commit does the same for Messages* files,
and Language subclasses (done separately for ease of review).

Change-Id: I301f471f86ba2dee924fece29a16dc3c20b5bebe