Merge away `mediawiki.ForeignStructuredUpload` module
Only needed as a dependency of
`mediawiki.ForeignStructuredUpload.BookletLayout`, can be combined.
To ensure that the upload code loads before the layout code, and
to maintain package file support for the configuration used in
ForeignStructuredUpload, convert booklet layout to package files
and add a basic entry point that just require()s the two files in order.

Bug: T311189
Change-Id: I8745125e4696acf232d56af119c0473a3fd8079d