Fix MemcachedPhpBagOStuff::merge() on a non-existent key
merge() on a non-existent key always fails because in all commands, the
client flags an empty result list as an error, causing mergeViaCas() to
abort with a log warning. This showed up as a test failure in
BagOStuffTestBase::testMerge(). It's unclear how that test could have
ever worked.

So, make it not do that. Follow MemcachedPeclBagOStuff in making "not
found" return false without being an error.

Change-Id: I0f70ce8acea00576d4bf79068d039f28111286f7