objectcache: Implement asynchronous increment in memcached BagOStuff
In incrWithInit(), when the WRITE_BACKGROUND flag is set, return
true on success, don't try to return the new value. Use the noreply
option to avoid waiting for a response.

assertEquals(3, true) unexpectedly succeeds, so use assertSame() in the
relevant tests. Fix the type of the incr() return value in
MemcachedClient, was string.

Bug: T310662
Change-Id: I1a41255c304b1ac2b6086e7820390c3cbad2e789